chore : 계약형태 컴포넌트 분리
